Generalized ftplugin dir; Changed '^M' to "<CR>"
Changed a couple of lines in lua_omni.vim so that it looks
in the directory in which it resides, wherever that may be, for
the lua_omni.lua file to be sourced, rather than
"~/.vim/ftplugin" specifically. This allows the user
to keep the plugin in a non-standard runtimepath location.

I also replaced the occurrences of '^M' in the mappings with the
more friendly (in my opinion) "<CR>". The behavior is, I believe,
identical, but it is easier to work with printable characters
(For example, I'm not sure how you would to a search for '^M'
occurrences in the file). An alternative could be "<C-M>".